ronund  -> roundf

+++ CPP-Module-02/ex01/Fixed.cpp
@@ -16,7 +16,7 @@ Fixed::Fixed(const int value)
 
 Fixed::Fixed(const float value)
 {
-	_value = round(value * (1 << _bits));
+	_value = roundf(value * (1 << _bits));
 	std::cout << "Float constructor called" << std::endl;
 }
